Refactor ParkingZone management: removed totalCapacity from ParkingZone and related DTOs, updated create and update methods in ZoneController and ZoneService to reflect changes, and adjusted data seeding accordingly.

+++ src/main/java/com/parking/zonemgmt/ParkingZoneDTO.java
@@ -7,7 +7,7 @@ public record ParkingZoneDTO(
         UUID id,
         String name,
         String address,
-        int totalCapacity,
+        long totalCapacity,
         long availableCount,
         Double latitude,
         Double longitude,
